Move agent.py to own dir
